requirements.txt
chromadb
fiona
geoalchemy2
geopandas
numpy
openai
openpyxl
psycopg2-binary
pyarrow
pyogrio
pyyaml
scikit-learn
sentence-transformers
sqlalchemy
torch
tqdm
transformers
